app/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
black==22.1.0